Sección de Políticas de acceso

+++ app/Http/Controllers/PostController.php
@@ -20,7 +20,9 @@ class PostController extends Controller
         return back()->with('status','Publicación guardada');
     }
 
-    public function destroy(Post $post) {
+    public function destroy(Request $request,Post $post) {
+
+        $this->authorize('delete', $post);
         $post->delete();
 
         return back();
